gradeDistribution
There are three possible ways to evaluate a grade
Option 1:
Midterm 1: 25% of the grade
Midterm 2: 25% of the grade
Final exam: 50% of the grade
Option 2:
Midterm 1: 50% of the grade
Midterm 2: 50% of the grade
Option 3:
Final exam: 100% of the grade

Task 1.  query the name and id of all the students whose best grade comes from Option 3
Task 2. sort based on the first 3 characters of their name
Task 3. If the first 3 characters of two names are the same, then the student with the lower ID value comes first